Reduce vine and leaf data

+++ butterflowers/Assets/Scripts/butterflowersOS/Data/SequenceData.cs
@@ -5,7 +5,7 @@ namespace butterflowersOS.Data
 	[System.Serializable]
 	public class SequenceData
 	{
-		public int index = -1;
+		public sbyte index = -1;
 		public Frame[] frames = new Frame[7];
 	}
 }
